moved loginfunction from apiFunctions to Views, easier to maintain
